removed LayerdBloomFilter.fixed()  methods
